Fix a delay in game while crossing a terrain and playing MIDI music (#5535)

+++ src/fheroes2/game/game.cpp
@@ -341,7 +341,7 @@ uint32_t & Game::MapsAnimationFrame()
 // play environment sounds from the game area in focus
 void Game::EnvironmentSoundMixer()
 {
-    size_t availableChannels = Mixer::getChannelCount();
+    int availableChannels = Mixer::getChannelCount();
     if ( availableChannels <= 2 ) {
         // 2 channels are left for hero's movement.
         return;
